Passcode Lock
Added passcode lock functionality. Use AsyncStorage to store the passcode. When user enable passcode lock, the user is required to enter passcode when open the app. Else, no need to enter passcode. (-- For the following modification, modify on the MainApp components in App.tsx, no more App components in App.tsx, because App components is now handle the passcode lock)